# webui/public/swagger/authenticated.json
{"swagger":"2.0","info":{"version":"0.5.0.0 (903a8d5)","title":"PostgREST API","description":"This is a dynamic API generated by PostgREST"},"host":"localhost:8180","basePath":"/","schemes":["http"],"consumes":["application/json","application/vnd.pgrst.object+json","text/csv"],"produces":["application/json","application/vnd.pgrst.object+json","text/csv"],"paths":{"/":{"get":{"tags":["Introspection"],"summary":"OpenAPI description (this document)","produces":["application/openapi+json","application/json"],"responses":{"200":{"description":"OK"}}}},"/comments":{"get":{"tags":["comments"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/comments"},"description":"OK"}}},"post":{"tags":["comments"],"parameters":[{"$ref":"#/parameters/body.comments"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["comments"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["comments"],"parameters":[{"$ref":"#/parameters/body.comments"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/files":{"get":{"tags":["files"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/files"},"description":"OK"}}},"post":{"tags":["files"],"parameters":[{"$ref":"#/parameters/body.files"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["files"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["files"],"parameters":[{"$ref":"#/parameters/body.files"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/ipsative_answers":{"get":{"tags":["ipsative_answers"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_answers"},"description":"OK"}}},"post":{"tags":["ipsative_answers"],"parameters":[{"$ref":"#/parameters/body.ipsative_answers"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["ipsative_answers"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["ipsative_answers"],"parameters":[{"$ref":"#/parameters/body.ipsative_answers"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/ipsative_categories":{"get":{"tags":["ipsative_categories"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_categories"},"description":"OK"}}},"post":{"tags":["ipsative_categories"],"parameters":[{"$ref":"#/parameters/body.ipsative_categories"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["ipsative_categories"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["ipsative_categories"],"parameters":[{"$ref":"#/parameters/body.ipsative_categories"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/ipsative_data":{"get":{"tags":["ipsative_data"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_data"},"description":"OK"}}}},"/ipsative_questions":{"get":{"tags":["ipsative_questions"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_questions"},"description":"OK"}}},"post":{"tags":["ipsative_questions"],"parameters":[{"$ref":"#/parameters/body.ipsative_questions"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["ipsative_questions"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["ipsative_questions"],"parameters":[{"$ref":"#/parameters/body.ipsative_questions"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/ipsative_responses":{"get":{"tags":["ipsative_responses"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_responses"},"description":"OK"}}},"post":{"tags":["ipsative_responses"],"parameters":[{"$ref":"#/parameters/body.ipsative_responses"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["ipsative_responses"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["ipsative_responses"],"parameters":[{"$ref":"#/parameters/body.ipsative_responses"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/ipsative_responses_grouped":{"get":{"tags":["ipsative_responses_grouped"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_responses_grouped"},"description":"OK"}}}},"/ipsative_surveys":{"get":{"tags":["ipsative_surveys"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/ipsative_surveys"},"description":"OK"}}},"post":{"tags":["ipsative_surveys"],"parameters":[{"$ref":"#/parameters/body.ipsative_surveys"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["ipsative_surveys"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["ipsative_surveys"],"parameters":[{"$ref":"#/parameters/body.ipsative_surveys"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_answers":{"get":{"tags":["likert_answers"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_answers"},"description":"OK"}}},"post":{"tags":["likert_answers"],"parameters":[{"$ref":"#/parameters/body.likert_answers"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_answers"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_answers"],"parameters":[{"$ref":"#/parameters/body.likert_answers"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_choice_groups":{"get":{"tags":["likert_choice_groups"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_choice_groups"},"description":"OK"}}},"post":{"tags":["likert_choice_groups"],"parameters":[{"$ref":"#/parameters/body.likert_choice_groups"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_choice_groups"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_choice_groups"],"parameters":[{"$ref":"#/parameters/body.likert_choice_groups"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_choices":{"get":{"tags":["likert_choices"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_choices"},"description":"OK"}}},"post":{"tags":["likert_choices"],"parameters":[{"$ref":"#/parameters/body.likert_choices"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_choices"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_choices"],"parameters":[{"$ref":"#/parameters/body.likert_choices"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_data":{"get":{"tags":["likert_data"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_data"},"description":"OK"}}}},"/likert_distinct_choice_groups":{"get":{"tags":["likert_distinct_choice_groups"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_distinct_choice_groups"},"description":"OK"}}}},"/likert_questions":{"get":{"tags":["likert_questions"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_questions"},"description":"OK"}}},"post":{"tags":["likert_questions"],"parameters":[{"$ref":"#/parameters/body.likert_questions"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_questions"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_questions"],"parameters":[{"$ref":"#/parameters/body.likert_questions"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_responses":{"get":{"tags":["likert_responses"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_responses"},"description":"OK"}}},"post":{"tags":["likert_responses"],"parameters":[{"$ref":"#/parameters/body.likert_responses"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_responses"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_responses"],"parameters":[{"$ref":"#/parameters/body.likert_responses"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}},"/likert_surveys":{"get":{"tags":["likert_surveys"],"parameters":[{"$ref":"#/parameters/select"},{"$ref":"#/parameters/order"},{"$ref":"#/parameters/range"},{"$ref":"#/parameters/rangeUnit"},{"$ref":"#/parameters/offset"},{"$ref":"#/parameters/limit"},{"$ref":"#/parameters/preferCount"}],"responses":{"206":{"description":"Partial Content"},"200":{"schema":{"$ref":"#/definitions/likert_surveys"},"description":"OK"}}},"post":{"tags":["likert_surveys"],"parameters":[{"$ref":"#/parameters/body.likert_surveys"},{"$ref":"#/parameters/preferReturn"}],"responses":{"201":{"description":"Created"}}},"delete":{"tags":["likert_surveys"],"parameters":[{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}},"patch":{"tags":["likert_surveys"],"parameters":[{"$ref":"#/parameters/body.likert_surveys"},{"$ref":"#/parameters/preferReturn"}],"responses":{"204":{"description":"No Content"}}}}},"definitions":{"comments":{"type":"object"},"files":{"type":"object"},"ipsative_answers":{"type":"object"},"ipsative_categories":{"type":"object"},"ipsative_data":{"type":"object"},"ipsative_questions":{"type":"object"},"ipsative_responses":{"type":"object"},"ipsative_responses_grouped":{"type":"object"},"ipsative_surveys":{"type":"object"},"likert_answers":{"type":"object"},"likert_choice_groups":{"type":"object"},"likert_choices":{"type":"object"},"likert_data":{"type":"object"},"likert_distinct_choice_groups":{"type":"object"},"likert_questions":{"type":"object"},"likert_responses":{"type":"object"},"likert_surveys":{"type":"object"}},"parameters":{"preferParams":{"name":"Prefer","description":"Preference","required":false,"in":"header","type":"string","enum":["params=single-object"]},"preferReturn":{"name":"Prefer","description":"Preference","required":false,"in":"header","type":"string","enum":["return=representation","return=minimal","return=none"]},"preferCount":{"name":"Prefer","description":"Preference","required":false,"in":"header","type":"string","enum":["count=none"]},"select":{"name":"select","description":"Filtering Columns","required":false,"in":"query","type":"string"},"order":{"name":"order","description":"Ordering","required":false,"in":"query","type":"string"},"range":{"name":"Range","description":"Limiting and Pagination","required":false,"in":"header","type":"string"},"rangeUnit":{"name":"Range-Unit","description":"Limiting and Pagination","required":false,"default":"items","in":"header","type":"string"},"offset":{"name":"offset","description":"Limiting and Pagination","required":false,"in":"query","type":"string"},"limit":{"name":"limit","description":"Limiting and Pagination","required":false,"in":"query","type":"string"},"body.comments":{"name":"comments","description":"comments","required":false,"schema":{"$ref":"#/definitions/comments"},"in":"body"},"body.files":{"name":"files","description":"files","required":false,"schema":{"$ref":"#/definitions/files"},"in":"body"},"body.ipsative_answers":{"name":"ipsative_answers","description":"ipsative_answers","required":false,"schema":{"$ref":"#/definitions/ipsative_answers"},"in":"body"},"body.ipsative_categories":{"name":"ipsative_categories","description":"ipsative_categories","required":false,"schema":{"$ref":"#/definitions/ipsative_categories"},"in":"body"},"body.ipsative_data":{"name":"ipsative_data","description":"ipsative_data","required":false,"schema":{"$ref":"#/definitions/ipsative_data"},"in":"body"},"body.ipsative_questions":{"name":"ipsative_questions","description":"ipsative_questions","required":false,"schema":{"$ref":"#/definitions/ipsative_questions"},"in":"body"},"body.ipsative_responses":{"name":"ipsative_responses","description":"ipsative_responses","required":false,"schema":{"$ref":"#/definitions/ipsative_responses"},"in":"body"},"body.ipsative_responses_grouped":{"name":"ipsative_responses_grouped","description":"ipsative_responses_grouped","required":false,"schema":{"$ref":"#/definitions/ipsative_responses_grouped"},"in":"body"},"body.ipsative_surveys":{"name":"ipsative_surveys","description":"ipsative_surveys","required":false,"schema":{"$ref":"#/definitions/ipsative_surveys"},"in":"body"},"body.likert_answers":{"name":"likert_answers","description":"likert_answers","required":false,"schema":{"$ref":"#/definitions/likert_answers"},"in":"body"},"body.likert_choice_groups":{"name":"likert_choice_groups","description":"likert_choice_groups","required":false,"schema":{"$ref":"#/definitions/likert_choice_groups"},"in":"body"},"body.likert_choices":{"name":"likert_choices","description":"likert_choices","required":false,"schema":{"$ref":"#/definitions/likert_choices"},"in":"body"},"body.likert_data":{"name":"likert_data","description":"likert_data","required":false,"schema":{"$ref":"#/definitions/likert_data"},"in":"body"},"body.likert_distinct_choice_groups":{"name":"likert_distinct_choice_groups","description":"likert_distinct_choice_groups","required":false,"schema":{"$ref":"#/definitions/likert_distinct_choice_groups"},"in":"body"},"body.likert_questions":{"name":"likert_questions","description":"likert_questions","required":false,"schema":{"$ref":"#/definitions/likert_questions"},"in":"body"},"body.likert_responses":{"name":"likert_responses","description":"likert_responses","required":false,"schema":{"$ref":"#/definitions/likert_responses"},"in":"body"},"body.likert_surveys":{"name":"likert_surveys","description":"likert_surveys","required":false,"schema":{"$ref":"#/definitions/likert_surveys"},"in":"body"}},"externalDocs":{"url":"https://postgrest.org/en/v5.0/api.html","description":"PostgREST Documentation"}}